Get rid of inner classes and some related shenanigans
