+++ task/src/main/java/pro/javacard/ant/JCImport.java
@@ -1,0 +1,15 @@
+package pro.javacard.ant;
+
+// Just for Ant
+public class JCImport {
+    String exps = null;
+    String jar = null;
+
+    public void setExps(String msg) {
+        exps = msg;
+    }
+
+    public void setJar(String msg) {
+        jar = msg;
+    }
+}
